Update buffer len variable types

+++ include/libsettings/settings_util.h
@@ -34,7 +34,7 @@ int settings_format(const char *section,
                     const char *value,
                     const char *type,
                     char *buf,
-                    uint8_t blen);
+                    uint16_t blen);
 
 /**
  * @brief   Parse setting strings from SBP message buffer
@@ -55,7 +55,7 @@ int settings_format(const char *section,
  * @retval -1               An error occurred.
  */
 settings_tokens_t settings_parse(const char *buf,
-                                 uint8_t blen,
+                                 uint16_t blen,
                                  const char **section,
                                  const char **name,
                                  const char **value,
